+++ FindPanel/OgreTextFindComponents/OgreTextFindBranch.m
@@ -188,12 +188,12 @@
     _isParentRetained = NO;
 }
 
-- (int)index
+- (NSInteger)index
 {
     return _index;
 }
 
-- (void)setIndex:(int)index
+- (void)setIndex:(NSInteger)index
 {
     _index = index;
 }
