+++ FindPanelTest/MyTableColumnSheet.m
@@ -44,7 +44,7 @@
 		contextInfo:nil];
 }
 
-- (void)sheetDidEnd:(NSWindow*)sheet returnCode:(int)returnCode contextInfo:(void*)contextInfo
+- (void)sheetDidEnd:(NSWindow*)sheet returnCode:(NSInteger)returnCode contextInfo:(void*)contextInfo
 {
 	[self release];
 }
